Better output of AIMSUN data.

+++ dev/Basic/geospatial/RoadSegment.hpp
@@ -54,7 +54,8 @@ public:
 
 public:
 	///Maximum speed of this road segment.
-	unsigned int maxSpeed;
+	//NOTE: I'm importing AIMSUN data which has speed as a double. Will need to consider why later. ~Seth
+	double maxSpeed;
 
 	void* roadBulge; ///<Currently haven't decided how to represent bulge.
